feat: change handle -> appId

+++ rust-packages/http-server/src/types.rs
@@ -1,3 +1,3 @@
-pub type AppId = u64;
+pub type AppId = u8;
 
 pub type HandleRequestCallback = extern "C" fn(*mut u8, u16, *mut u8);
